Add debugging option and print driver status
Added option to debug the code by changing default_env from release to
debug in platformio.ini. Debugging port must match the port the PCB is
connected to. Debugging is started by going to Run->Start Debugging.

Release mode is set by chaning default_envs to release. It is started by
going to Platformio->release->Upload and Monitor.
In release mode state of drivers is updated every 10s and STATUS1 to
STATUS5 register values are printed to the terminal.

KNOWN ISSUE: in debug mode UVLO3 fault bit in STATUS3 is set correctly
when no voltage is present on VEE, but in release mode the fault bit is
not set